Implemented inital keyboard input.

+++ code/handmade.cpp
@@ -38,8 +38,8 @@ RenderWeirdGradient(game_offscreen_buffer* Buffer, int BlueOffset, int GreenOffs
 			++X)
 		{
 			// This is BLUE
-			uint8 Blue = (X + BlueOffset);
-			uint8 Green = (Y + GreenOffset);
+			uint8 Blue = (uint8)(X + BlueOffset);
+			uint8 Green = (uint8)(Y + GreenOffset);
 
 			/*
 				Memory:		BB GG RR xx
@@ -81,7 +81,7 @@ GameUpdateAndRender(game_memory* Memory, game_input* Input, game_offscreen_buffe
 	if (Input0->IsAnalog)
 	{
 		// NOTE: Use analog movement tuning
-		GameState->BlueOffset += (int)4.0f * (Input0->EndX);
+		GameState->BlueOffset += (int)(4.0f * (Input0->EndX));
 		GameState->ToneHz = 256 + (int)(128.0f * (Input0->EndY));
 	}
 	else
